fix: _webapi_input gets rewritten to native function after being called
I feel like this is new, because this used to work. Something is rewriting _webapi_input to the native function after we call it. What I’ve done is a hack to change it back. It’s not the `withStackSave` stuff… it still does it with my original blander code.